Moved readfile() shim to its own trait
